Refactoring the core part of alpha-fsm module to move to alpha-core
