+++ alpha/alpha-fsm/src/test/java/org/apache/servicecomb/pack/alpha/fsm/RedisChannelTest.java
@@ -20,9 +20,9 @@ import org.apache.servicecomb.pack.alpha.core.NodeStatus;
 import org.apache.servicecomb.pack.alpha.fsm.channel.redis.MessageSerializer;
 import org.apache.servicecomb.pack.alpha.fsm.channel.redis.RedisMessagePublisher;
 import org.apache.servicecomb.pack.alpha.fsm.channel.redis.RedisMessageSubscriber;
-import org.apache.servicecomb.pack.alpha.fsm.event.base.BaseEvent;
+import org.apache.servicecomb.pack.alpha.core.fsm.event.base.BaseEvent;
 import org.apache.servicecomb.pack.alpha.fsm.metrics.MetricsService;
-import org.apache.servicecomb.pack.alpha.fsm.sink.ActorEventSink;
+import org.apache.servicecomb.pack.alpha.core.fsm.sink.ActorEventSink;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
